+++ src/components/spartan/SModal/SModalTitle.vue
@@ -1,0 +1,7 @@
+<script setup lang="ts">
+import { DialogTitle } from '@headlessui/vue';
+</script>
+
+<template>
+    <DialogTitle><slot /></DialogTitle>
+</template>