+++ src/components/spartan/SModal/index.ts
@@ -1,0 +1,3 @@
+export { default as SModal } from './SModal.vue';
+export { default as SModalTitle } from './SModalTitle.vue';
+export { default as SModalDescription } from './SModalDescription.vue';