+++ src/components/spartan/index.ts
@@ -21,3 +21,4 @@ export * from './SDefinitionTerm';
 export * from './SToast';
 export * from './SSteps';
 export * from './SBreadcrumbs';
+export * from './SModal';